moving functions to more coherent locations
